+++ data/src/main/java/dog/shebang/data/RemoteMetadataDataSource.kt
@@ -1,15 +1,10 @@
 package dog.shebang.data
 
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
 import dog.shebang.data.api.LinkPreviewApiClient
 import dog.shebang.model.Metadata
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
-import javax.inject.Singleton
 
 interface RemoteMetadataDataSource {
 
@@ -27,16 +22,4 @@ class RemoteMetadataDataSourceImpl @Inject constructor(
             emit(null)
         }
     }
-}
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class RemoteMetadataDataSourceModule {
-
-    @Singleton
-    @Binds
-    abstract fun bindRemoteMetadataDataSource(
-        remoteMetadataDataSourceImpl: RemoteMetadataDataSourceImpl
-    ): RemoteMetadataDataSource
-
 }
