+++ data/src/main/java/dog/shebang/data/di/LinkPreviewApiKeyModule.kt
@@ -4,8 +4,8 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import dog.shebang.env.Environment
-import javax.inject.Inject
+import dog.shebang.data.api.LinkPreviewApiKey
+import dog.shebang.data.api.LinkPreviewApiKeyImpl
 import javax.inject.Singleton
 
 @Module
@@ -18,17 +18,4 @@ abstract class LinkPreviewApiKeyModule {
         linkPreviewApiKeyImpl: LinkPreviewApiKeyImpl
     ): LinkPreviewApiKey
 
-}
-
-interface LinkPreviewApiKey {
-
-    val apiKey: String
-}
-
-class LinkPreviewApiKeyImpl @Inject constructor(
-    private val environment: Environment
-) : LinkPreviewApiKey {
-
-    override val apiKey: String
-        get() = environment.linkPreviewApiKey
 }
